requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.1
django-cors-headers==4.0.0
django-jazzmin==2.6.0
django-ratelimit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
rfc3986==2.0.0
sniffio==1.3.0
sqlparse==0.4.4
typing-extensions==4.6.2
whitenoise==6.4.0
drf-yasg==1.21.5 
inflection==0.5.1 
packaging==23.1 
ruamel.yaml==0.17.31 
ruamel.yaml.clib==0.2.7
decorator==5.1.1 
py==1.11.0 
retry==0.9.2
cron-descriptor==1.4.0 
django-celery-beat==2.5.0 
django-timezone-field==5.0 
python-crontab==2.7.1